+++ lib/tr_extension.dart
@@ -1,5 +1,6 @@
 library tr_extension;
 
+import 'dart:collection';
 import 'dart:convert';
 import 'dart:developer' as dev;
 
@@ -8,6 +9,7 @@ import 'package:flutter/services.dart';
 import 'package:flutter/widgets.dart';
 import 'package:flutter_localizations/flutter_localizations.dart';
 import 'package:intl/intl.dart';
+
 export 'package:intl/intl.dart';
 export 'package:tr_extension/tr_extension.dart';